Added ZCU106 QEMU config file and image directory
QEMU command lines can become rather lengthy, so the Xilinx config was
dumped to an ini style config file and can be read from QEMU using the
`-readconfig` command line arg, although not all options are supported
in the config file so the final launcher will include a mix.

Currently, users are expected to place the required images in the image
directory. They are currently ignored by git due to the large file
sizes.